backend/pyproject.toml
[project]
name = "goals"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "motor==3.5.1",
    "pymongo==4.8.0",
    "pydantic==2.8.2",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
